serv/requirements.txt
flask
python-dotenv
openai
langchain
tiktoken
faiss-cpu
unstructured
pandas
numpy